refactor: extract DB schema and format test file
SQL schemas + migrations → db/schema.rs
db/mod.rs keeps Database struct, open, conn, migrate only.
